Created Search box and App styles added
