+++ src/App.jsx
@@ -1,0 +1,12 @@
+import "./App.css";
+import SearchBox from "./components/search-box/search-box.component";
+
+function App() {
+  return (
+    <div>
+      <SearchBox />
+    </div>
+  );
+}
+
+export default App;
