+++ src/components/search-box/search-box.component.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import './search-box.styles.css'
+
+const SearchBox = () => {
+  return <input type="search" className="search-box" />;
+};
+
+export default SearchBox;
