+++ src/components/search-box/search-box.styles.css
@@ -1,0 +1,11 @@
+.search-box {
+  --webkit-appearance: none;
+  border: none;
+  outline: none;
+  padding: 10px;
+  width: 25%;
+  font-size: large;
+  line-height: 30px;
+  margin-bottom: 30px;
+  border-radius: 6px;
+}
